requirements.txt
pandas==2.0.3
numpy==1.24.4
matplotlib==3.7.5
plotly==5.24.1
streamlit==1.29.0
python-dotenv==1.0.0
